Stop using Replicator interface in bootstrap and heathcheck binaries.
